+++ FlinkDotNet/FlinkDotNet.Storage.FileSystem.Tests/FileSystemSnapshotReaderTests.cs
@@ -14,8 +14,26 @@ namespace FlinkDotNet.Storage.FileSystem.Tests
         }
 
         public void Dispose()
+
+
         {
-            if (Directory.Exists(_testDirectory))
+
+
+            Dispose(true);
+
+
+            GC.SuppressFinalize(this);
+
+
+        }
+
+
+
+        protected virtual void Dispose(bool disposing)
+
+
+        {
+            if (disposing && Directory.Exists(_testDirectory))
             {
                 Directory.Delete(_testDirectory, recursive: true);
             }
@@ -362,6 +380,9 @@ namespace FlinkDotNet.Storage.FileSystem.Tests
 
             // Act & Assert
             reader.Dispose(); // Should not throw
+            
+            // Assert - verify reader is in expected state after disposal
+            Assert.NotNull(reader);
         }
     }
 }
