+++ FlinkDotNet/FlinkDotNet.Storage.FileSystem.Tests/FileSystemSnapshotStoreTests.cs
@@ -15,7 +15,13 @@ namespace FlinkDotNet.Storage.FileSystem.Tests
 
         public void Dispose()
         {
-            if (Directory.Exists(_testDirectory))
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+            if (disposing && Directory.Exists(_testDirectory))
             {
                 Directory.Delete(_testDirectory, recursive: true);
             }
@@ -30,7 +36,7 @@ namespace FlinkDotNet.Storage.FileSystem.Tests
             try
             {
                 // Act
-                var store = new FileSystemSnapshotStore(tempPath);
+                new FileSystemSnapshotStore(tempPath);
 
                 // Assert
                 Assert.True(Directory.Exists(tempPath));
@@ -54,7 +60,7 @@ namespace FlinkDotNet.Storage.FileSystem.Tests
             try
             {
                 // Act
-                var store = new FileSystemSnapshotStore(relativePath);
+                new FileSystemSnapshotStore(relativePath);
 
                 // Assert
                 Assert.True(Directory.Exists(expectedPath));
