+++ FlinkDotNet/FlinkDotNet.Storage.FileSystem.Tests/FileSystemSnapshotWriterTests.cs
@@ -13,13 +13,18 @@ namespace FlinkDotNet.Storage.FileSystem.Tests
 
         public void Dispose()
         {
-            if (Directory.Exists(_testDirectory))
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+            if (disposing && Directory.Exists(_testDirectory))
             {
                 try
                 {
-                    // Force garbage collection to close any lingering file handles
-                    GC.Collect();
-                    GC.WaitForPendingFinalizers();
+                    // Wait briefly to allow file handles to be released
+                    Thread.Sleep(10);
                     
                     Directory.Delete(_testDirectory, recursive: true);
                 }
